requirements.txt
# Discord Bot Framework
discord.py[voice]>=2.3.0

# Environment Variables
python-dotenv>=1.0.0

# Database
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0

# Video/Audio Downloading
yt-dlp>=2023.0.0

# Curl
curl_cffi>=2.0.0

# JWT for Authentication
PyJWT>=2.8.0

# HTTP Client
aiohttp>=3.11.0

# RSS Feed
feedparser>=6.0.0
